refactor(api): extract public hunk-manipulation API and expand hooks
Previously all core logic (toggle_file, toggle_lines, toggle_hunk,
open_file, set_global_bindings) lived as private local functions inside
lua/hunk/init.lua, and a module-level mutable CONTEXT global held
session state. Users could not call any of these functions directly —
the only customization surface was swapping key chord strings.

This commit:

- Adds lua/hunk/api/actions.lua as the public API module, exposing
  toggle_file, toggle_lines, toggle_line_pairs, toggle_hunk, open_file,
  and set_global_bindings.
- Registers actions in lua/hunk/api/init.lua so users can call
  require("hunk.api").actions.toggle_lines(...) etc.
- Replaces the CONTEXT global with a first-class session object
  (hunk.Session) that is created per start() call and returned to the
  caller. All functions now receive session as an explicit parameter.
- Re-implements existing keybindings in init.lua through api.actions.*
  rather than private locals.
- Adds 9 new hooks to hunk.Config.hooks with LuaLS annotations:
  on_file_opened, on_file_previewed, on_file_toggled,
  on_line_toggled, on_hunk_toggled, on_hunk_navigated,
  on_focus_changed, on_before_accept, on_before_quit.
- Wires hook calls into tree.lua, file.lua, and actions.lua at the
  appropriate lifecycle points.
- Makes file.lua import only the sub-modules it needs (signs, fs)
  instead of the full api bundle.

No user-visible keybinding behavior changes. The default hooks are
all no-ops, preserving existing semantics.
